Reorg entity out of model
