+++ core/src/lib.rs
@@ -3,6 +3,7 @@ pub mod collation;
 pub mod comparison_index;
 pub mod connector;
 pub mod context;
+pub mod entity;
 pub mod error;
 pub mod event;
 pub mod model;
